backend/requirements.txt
Django==5.0.4
asgiref==3.8.1
sqlparse==0.4.4
tzdata==2024.1
djangorestframework==3.15.1
django-cors-headers==4.3.1

